requirements.txt
openai-agents>=0.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
